feat: Finish day5 pt 2

+++ cpp/src/day05.hpp
@@ -4,5 +4,5 @@
 
 namespace day05 {
 int part1(const std::vector<std::string> &list);
-int part2(const std::vector<std::string> &list);
+size_t part2(const std::vector<std::string> &list);
 } // namespace day05
